refactor(ParticipantDto): ParticipantVo -> ParticipantDto 변경
