+++ src/main/java/blackjack/domain/participant/Dealer.java
@@ -1,7 +1,5 @@
 package blackjack.domain.participant;
 
-import java.util.List;
-
 import blackjack.domain.PlayStatus;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardDeck;
